Make GCM receiver display notifications and move to own package
