+++ app/src/main/java/org/matrix/matrixandroidsdk/gcm/GcmBroadcastReceiver.java
@@ -14,15 +14,13 @@
  * limitations under the License.
  */
 
-package org.matrix.matrixandroidsdk;
+package org.matrix.matrixandroidsdk.gcm;
 
 import android.app.Activity;
 import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
 import android.support.v4.content.WakefulBroadcastReceiver;
-
-import org.matrix.matrixandroidsdk.services.GcmIntentService;
 
 /**
  * Receiver for GCM messages that starts GcmIntentService to handle the message.
